Create separate modules for API models and networking code
