+++ app/build.gradle
@@ -58,6 +58,8 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
+    implementation project(":data:api")
+    implementation project(":data:networking")
     implementation deps.androidX.coreKtx
     implementation deps.androidX.appCompat
     implementation deps.androidX.constraintLayout
