+++ app/src/main/java/com/alvindizon/tampisaw/core/ModelExt.kt
@@ -1,10 +1,10 @@
 package com.alvindizon.tampisaw.core
 
-import com.alvindizon.tampisaw.data.networking.model.getcollections.GetCollectionsResponse
-import com.alvindizon.tampisaw.data.networking.model.getphoto.GetPhotoResponse
-import com.alvindizon.tampisaw.data.networking.model.listphotos.ListPhotosResponse
-import com.alvindizon.tampisaw.data.networking.model.listphotos.Urls
-import com.alvindizon.tampisaw.data.networking.model.listphotos.User
+import com.alvindizon.tampisaw.api.model.getcollections.GetCollectionsResponse
+import com.alvindizon.tampisaw.api.model.getphoto.GetPhotoResponse
+import com.alvindizon.tampisaw.api.model.listphotos.ListPhotosResponse
+import com.alvindizon.tampisaw.api.model.listphotos.Urls
+import com.alvindizon.tampisaw.api.model.listphotos.User
 import com.alvindizon.tampisaw.features.collections.UnsplashCollection
 import com.alvindizon.tampisaw.features.details.PhotoDetails
 import com.alvindizon.tampisaw.features.gallery.UnsplashPhoto
@@ -32,25 +32,25 @@ fun GetPhotoResponse.toPhotoDetails() = PhotoDetails(
 fun GetPhotoResponse.getLocationString(): String = when {
     location.city != null && location.country != null ->
         "${location.city}, ${location.country}"
-    location.city != null && location.country == null -> location.city
-    location.city == null && location.country != null -> location.country
+    location.city != null && location.country == null -> location.city!!
+    location.city == null && location.country != null -> location.country!!
     else -> "N/A"
 }
 
 fun GetPhotoResponse.getCamera(): String {
     return exif.run {
-        model?.apply{
+        model?.apply {
             val makeList = make?.split(" ")?.map { it.trim().lowercase(Locale.ROOT) }
-            val modelList = model.split(" ").map { it.trim().lowercase(Locale.ROOT) }
+            val modelList = split(" ").map { it.trim().lowercase(Locale.ROOT) }
 
-            if(makeList?.intersect(modelList)?.isEmpty() == true) {
+            if (makeList?.intersect(modelList)?.isEmpty() == true) {
                 "${makeList.first()} $model"
             } else model
         }
     } ?: "Unknown"
 }
 
-fun GetPhotoResponse.getTags(): List<String?>?  = tags?.map { it.title }
+fun GetPhotoResponse.getTags(): List<String?>? = tags?.map { it.title }
 
 fun GetCollectionsResponse.toUnsplashCollection() = UnsplashCollection(
     id, title, description, user.name, user.username, user.profileImage?.large, isPrivate,
