+++ app/src/main/java/com/alvindizon/tampisaw/data/UnsplashRepoImpl.kt
@@ -4,11 +4,11 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import androidx.paging.rxjava3.observable
+import com.alvindizon.tampisaw.api.UnsplashApi
+import com.alvindizon.tampisaw.api.model.getcollections.GetCollectionsResponse
+import com.alvindizon.tampisaw.api.model.getphoto.GetPhotoResponse
+import com.alvindizon.tampisaw.api.model.listphotos.ListPhotosResponse
 import com.alvindizon.tampisaw.core.Const
-import com.alvindizon.tampisaw.data.networking.api.UnsplashApi
-import com.alvindizon.tampisaw.data.networking.model.getcollections.GetCollectionsResponse
-import com.alvindizon.tampisaw.data.networking.model.getphoto.GetPhotoResponse
-import com.alvindizon.tampisaw.data.networking.model.listphotos.ListPhotosResponse
 import com.alvindizon.tampisaw.data.paging.CollectionPagingSource
 import com.alvindizon.tampisaw.data.paging.SearchCollectionsPagingSource
 import com.alvindizon.tampisaw.data.paging.SearchPhotosPagingSource
