+++ app/src/main/java/com/alvindizon/tampisaw/data/paging/CollectionPagingSource.kt
@@ -2,9 +2,9 @@ package com.alvindizon.tampisaw.data.paging
 
 import androidx.paging.PagingState
 import androidx.paging.rxjava3.RxPagingSource
+import com.alvindizon.tampisaw.api.UnsplashApi
+import com.alvindizon.tampisaw.api.model.getcollections.GetCollectionsResponse
 import com.alvindizon.tampisaw.core.Const
-import com.alvindizon.tampisaw.data.networking.api.UnsplashApi
-import com.alvindizon.tampisaw.data.networking.model.getcollections.GetCollectionsResponse
 import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.schedulers.Schedulers
