+++ app/src/main/java/com/alvindizon/tampisaw/domain/UnsplashRepo.kt
@@ -1,9 +1,9 @@
 package com.alvindizon.tampisaw.domain
 
 import androidx.paging.PagingData
-import com.alvindizon.tampisaw.data.networking.model.getcollections.GetCollectionsResponse
-import com.alvindizon.tampisaw.data.networking.model.getphoto.GetPhotoResponse
-import com.alvindizon.tampisaw.data.networking.model.listphotos.ListPhotosResponse
+import com.alvindizon.tampisaw.api.model.getcollections.GetCollectionsResponse
+import com.alvindizon.tampisaw.api.model.getphoto.GetPhotoResponse
+import com.alvindizon.tampisaw.api.model.listphotos.ListPhotosResponse
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.core.Single
