+++ app/src/test/java/com/alvindizon/tampisaw/features/details/DetailsViewModelTest.kt
@@ -6,14 +6,14 @@ import androidx.lifecycle.LifecycleOwner
 import com.alvindizon.tampisaw.CoroutineExtension
 import com.alvindizon.tampisaw.InstantExecutorExtension
 import com.alvindizon.tampisaw.RxSchedulerExtension
+import com.alvindizon.tampisaw.api.model.getphoto.Exif
+import com.alvindizon.tampisaw.api.model.getphoto.GetPhotoResponse
+import com.alvindizon.tampisaw.api.model.getphoto.Links
+import com.alvindizon.tampisaw.api.model.getphoto.Location
+import com.alvindizon.tampisaw.api.model.getphoto.Tag
+import com.alvindizon.tampisaw.api.model.getphoto.Urls
+import com.alvindizon.tampisaw.api.model.getphoto.User
 import com.alvindizon.tampisaw.core.toPhotoDetails
-import com.alvindizon.tampisaw.data.networking.model.getphoto.Exif
-import com.alvindizon.tampisaw.data.networking.model.getphoto.GetPhotoResponse
-import com.alvindizon.tampisaw.data.networking.model.getphoto.Links
-import com.alvindizon.tampisaw.data.networking.model.getphoto.Location
-import com.alvindizon.tampisaw.data.networking.model.getphoto.Tag
-import com.alvindizon.tampisaw.data.networking.model.getphoto.Urls
-import com.alvindizon.tampisaw.data.networking.model.getphoto.User
 import com.alvindizon.tampisaw.domain.DownloadPhotoUseCase
 import com.alvindizon.tampisaw.domain.GetPhotoUseCase
 import com.alvindizon.tampisaw.domain.SetWallpaperByBitmapUseCase
