+++ data/api/src/main/java/com/alvindizon/tampisaw/api/model/search/SearchModels.kt
@@ -1,0 +1,26 @@
+package com.alvindizon.tampisaw.api.model.search
+
+import com.alvindizon.tampisaw.api.model.getcollections.GetCollectionsResponse
+import com.alvindizon.tampisaw.api.model.listphotos.ListPhotosResponse
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class SearchPhotosResponse(
+    @Json(name = "total")
+    val total: Int,
+    @Json(name = "total_pages")
+    val totalPages: Int,
+    @Json(name = "results")
+    val results: List<ListPhotosResponse>
+)
+
+@JsonClass(generateAdapter = true)
+data class SearchCollectionsResponse(
+    @Json(name = "total")
+    val total: Int,
+    @Json(name = "total_pages")
+    val totalPages: Int,
+    @Json(name = "results")
+    val results: List<GetCollectionsResponse>
+)
