+++ data/networking/src/main/java/com/alvindizon/tampisaw/networking/interceptor/ConnectivityInterceptor.kt
@@ -1,6 +1,6 @@
-package com.alvindizon.tampisaw.data.networking.interceptor
+package com.alvindizon.tampisaw.networking.interceptor
 
-import com.alvindizon.tampisaw.data.networking.exception.NoInternetException
+import com.alvindizon.tampisaw.networking.exception.NoInternetException
 import okhttp3.Interceptor
 import okhttp3.Response
 import java.io.IOException
@@ -8,7 +8,9 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class ConnectivityInterceptor @Inject constructor(private val monitor: LiveNetworkMonitor) : Interceptor {
+class ConnectivityInterceptor @Inject constructor(
+    private val monitor: LiveNetworkMonitor
+) : Interceptor {
 
     @Throws(IOException::class)
     override fun intercept(chain: Interceptor.Chain): Response {
