+++ settings.gradle
@@ -1,2 +1,4 @@
 include ':app'
 rootProject.name = "Tampisaw"
+include ':data:api'
+include ':data:networking'
